requirements.txt
deepface==0.0.96
fastapi==0.123.8
Jinja2==3.1.6
numpy==2.2.6
opencv-python==4.12.0.88
oracledb==3.4.0
pillow==12.0.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.1
python-multipart==0.0.20
scikit-learn==1.7.2
tf-keras==2.20.1
torch==2.9.0
torchvision==0.24.0
ultralytics==8.3.221
ultralytics-thop==2.0.17
uvicorn==0.38.0